break nn modules into files
